fix: the request for saving the actions works!
