+++ .devcontainer.json
@@ -1,20 +1,14 @@
 {
-	"name": "python-starter",
-	"image": "ghcr.io/nuvolaris/devcontainer:3.1.0-mastrogpt.2403171925",
-	"mounts": [
-		"source=/var/run/docker.sock,target=/var/run/docker-host.sock,type=bind"
-	],
-	"remoteUser": "nuvolaris",
-	"overrideCommand": false,
-	"runArgs": [
-		"--network=host"
-	],
-	"customizations": {
-		"vscode": {
-			"extensions": [
-				"nuvolaris.nuvolaris-vscode-extension@1.0.2",
-				"ms-python.python"
-			]
-		}
-	}
+    "name": "nuvolaris-starter",
+    "image": "ghcr.io/nuvolaris/devcontainer:3.1.0-mastrogpt.2404131430",
+    "runArgs": [
+        "--init"
+    ],
+    "customizations": {
+        "vscode": {
+            "extensions": [
+                "nuvolaris.nuvolaris-vscode-extension"
+            ]
+        }
+    }
 }
